Added gherkin tests for trainee profile page and restructure some things.
