+++ src/test/java/com/sparta/eng82/tests/stepdefs/stepdefutil/Pages.java
@@ -4,6 +4,7 @@ import com.sparta.eng82.components.pages.admin.AddTrainerPageImpl;
 import com.sparta.eng82.components.pages.admin.AdminHomePageImpl;
 import com.sparta.eng82.components.pages.other.LoginPageImpl;
 import com.sparta.eng82.components.pages.trainee.TraineeHomePageImpl;
+import com.sparta.eng82.components.pages.trainee.TraineeProfilePageImpl;
 import com.sparta.eng82.components.pages.trainer.TrainerHomePageImpl;
 
 public class Pages {
@@ -12,4 +13,5 @@ public class Pages {
     public static TrainerHomePageImpl trainerHomePage;
     public static TraineeHomePageImpl traineeHomePage;
     public static AddTrainerPageImpl addTrainerPage;
+    public static TraineeProfilePageImpl traineeProfilePage;
 }
