+++ src/test/java/com/sparta/eng82/tests/stepdefs/trainee/TraineeStepDefs.java
@@ -1,9 +1,10 @@
-package com.sparta.eng82.tests.stepdefs;
+package com.sparta.eng82.tests.stepdefs.trainee;
 
 import com.sparta.eng82.components.Page;
 import com.sparta.eng82.components.pages.admin.AdminHomePageImpl;
 import com.sparta.eng82.components.pages.other.LoginPageImpl;
 import com.sparta.eng82.components.pages.trainee.TraineeHomePageImpl;
+import com.sparta.eng82.components.pages.trainee.TraineeProfilePageImpl;
 import com.sparta.eng82.tests.stepdefs.stepdefutil.DriverManager;
 import com.sparta.eng82.tests.stepdefs.stepdefutil.Pages;
 import io.cucumber.java.After;
@@ -13,7 +14,7 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.junit.jupiter.api.Assertions;
 
-public class TraineeHomepageStepDefs {
+public class TraineeStepDefs {
 
 
     @Given("I am logged in as a trainee")
@@ -45,5 +46,15 @@ public class TraineeHomepageStepDefs {
     @Then("I should see a selectable list of all the 1-1 sheets from previous weeks")
     public void iShouldSeeASelectableListOfAllTheSheetsFromPreviousWeeks() {
         Assertions.assertTrue(Pages.traineeHomePage.areAllPreviousWeeksShown());
+    }
+
+    @When("The trainee views their profile")
+    public void theTraineeViewsTheirProfile() {
+        Pages.traineeProfilePage = (TraineeProfilePageImpl) Pages.traineeHomePage.goToProfilePage();
+    }
+
+    @Then("The trainee will be able to see their average grades over time")
+    public void theTraineeWillBeAbleToSeeTheirAverageGradesOverTime() {
+        Assertions.assertTrue(Pages.traineeProfilePage.areBothGradesValid());
     }
 }
