--- src/test/resources/features/trainee/userStory-63.feature
@@ -1,7 +1,0 @@
-Scenario: The Trainee wants to view feedback progress over time
-
-Given The Trainee is logged in
-
-When The Trainee views their profile
-
-Then The Trainee will able to see their average grades over time.
